[core] Fix task log info fields in lifecycle events (#65190)
Found the bug when working on task event migration from GCS to dashboard
head. I switched off the task event buffer to GCS direct flag and rather
routed task events via the aggregator. This led to failing
`test_log_get` and `test_log_task` in `test_state_api_log` due to this
error:

`ray.util.state.exception.RayStateApiException: Could not find log file
for task attempt:`

`ray logs task <task_id>` resolves a task's log file using
`task_log_info`. When task events reach GCS through the event aggregator
(via the ray event framework), the field never arrives, and the lookup
fails. This happens because the `gcs_ray_event_converter` never maps
task log info to `TaskEvents` before storing them. Since we want to
store them the right way, we want to check if fields were ever populated
by source. Since we need `has_*` checks for them, we also add optional
to the fields in the proto `TaskLifecycleEvent`.

---------

+++ src/ray/common/protobuf_utils.cc
@@ -380,10 +380,12 @@ void TaskLogInfoToExport(const rpc::TaskLogInfo &src,
                          rpc::ExportTaskEventData::TaskLogInfo *dest) {
   dest->set_stdout_file(src.stdout_file());
   dest->set_stderr_file(src.stderr_file());
-  dest->set_stdout_start(src.stdout_start());
-  dest->set_stdout_end(src.stdout_end());
-  dest->set_stderr_start(src.stderr_start());
-  dest->set_stderr_end(src.stderr_end());
+  // The export schema is public and its offsets are int32, so the wider offsets are
+  // narrowed here rather than changing a field type that consumers already parse.
+  dest->set_stdout_start(static_cast<int32_t>(src.stdout_start()));
+  dest->set_stdout_end(static_cast<int32_t>(src.stdout_end()));
+  dest->set_stderr_start(static_cast<int32_t>(src.stderr_start()));
+  dest->set_stderr_end(static_cast<int32_t>(src.stderr_end()));
 }
 
 std::optional<rpc::autoscaler::PlacementConstraint>
